Update Aquarium to handle lost context
I'm not actually sure this is correct. The issue the code asks for meshes
and textures to be loaded async. So say the steps are this

request texture to be loaded #1a
context lost
context retored
request texture to be loaded #1b

texture request #1a completes
texture request #1b completes

Now the same texture has been loaded twice into the new context.
For textures, I could try to handle this in TDL, just don't
call the callbacks. But that won't work because user callbacks
need to be called. For scenes that needs to be handled here.